requirements.txt
Django>=3.2.7,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
python-dotenv>=0.19.0,<0.20.0

flake8>=3.9.2,<3.10.0

